Attempt supporting spaces in CC path
Autoconf does not support spaces in the compiler path. This is not
usually a problem, as on most platforms, the compiler will reside in a
path without spaces (such as /usr/bin or /opt). However, on windows, the
MSVC compiler is usually found under C:\Program Files, which causes all
hell to break loose.

Thankfully, windows has something called "Short Paths", which we can use
to turn C:\Program Files into C:\PROGRA~2, eliminating the path.

If the build script detects the compiler path has a space, and is
running on windows, it will attempt to turn the path into a short path
and run with this.